% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dateline.R
\name{remove_dateline}
\alias{remove_dateline}
\title{remove_dateline}
\usage{
remove_dateline(x)
}
\arguments{
\item{x}{The object from which the dateline should be removed}
}
\value{

}
\description{
This function removes the annoying dateline artifact from sf objects that cross the dateline
Code: x%>%st_buffer(1) %>%st_transform(st_crs(3338)) %>%st_simplify
}
\examples{
x<-remove_dateline(x)
}
